# Conflicts:
#	README.md
#	USER_GUIDE.md
#	docs/sd-card-fonts.md
#	docs/translators.md
#	docs/troubleshooting.md
#	docs/webserver-endpoints.md
#	docs/webserver.md
#	lib/Epub/Epub/parsers/ChapterHtmlSlimParser.cpp
#	lib/I18n/translations/english.yaml
#	src/activities/network/WifiSelectionActivity.cpp
#	src/activities/reader/ReaderLinkPickerActivity.cpp
#	src/network/html/SettingsPage.html

+++ src/activities/home/FileBrowserActivity.h
@@ -20,7 +20,7 @@ class FileBrowserActivity final : public Activity {
 
   ButtonNavigator buttonNavigator;
 
-  size_t selectorIndex = 0;
+  uint16_t selectorIndex = 0;
 
   bool lockLongPressBack = false;
   // True when this activity was entered while Confirm was already held; we must swallow the next
@@ -36,7 +36,7 @@ class FileBrowserActivity final : public Activity {
 
   // Data loading
   void loadFiles();
-  size_t findEntry(const std::string& name) const;
+  uint16_t findEntry(const std::string& name) const;
 
  public:
   explicit FileBrowserActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, std::string initialPath = "/",
