+++ src/activities/settings/StatusBarSettingsActivity.h
@@ -1,7 +1,4 @@
 #pragma once
-#include <freertos/FreeRTOS.h>
-#include <freertos/semphr.h>
-#include <freertos/task.h>
 
 #include "activities/Activity.h"
 #include "util/ButtonNavigator.h"
@@ -20,9 +17,9 @@ class StatusBarSettingsActivity final : public Activity {
  private:
   ButtonNavigator buttonNavigator;
 
-  int selectedIndex = 0;
+  uint8_t selectedIndex = 0;
   // Decided in onEnter() based on halClock.isAvailable() so clock entries are hidden on X4.
-  int visibleItemCount = 0;
+  uint8_t visibleItemCount = 0;
 
   void handleSelection();
 };
